Don't try to make a cast value an lvalue (Brian Johnson). Add some explicit casts to (int) in order to avaoid this warning: warning: comparison between `const enum video_depth' and `enum <anonymous>'

+++ BasiliskII/src/Unix/video_blit.h
@@ -69,7 +69,7 @@ enum {
 #define VIDEO_MODE_X			mode.x
 #define VIDEO_MODE_Y			mode.y
 #define VIDEO_MODE_RESOLUTION	mode.resolution_id
-#define VIDEO_MODE_DEPTH		(int)mode.depth
+#define VIDEO_MODE_DEPTH		mode.depth
 #endif
 
 #endif /* VIDEO_BLIT_H */
